Issue #80 - new ExceptionGenerator classes added, Exception handling moved to these new classes
