+++ source/library/com/restfb/exception/generator/LegacyFacebookExceptionGenerator.java
@@ -1,0 +1,49 @@
+/**
+ * Copyright (c) 2010-2015 Mark Allen, Norbert Bartels.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package com.restfb.exception.generator;
+
+import com.restfb.exception.FacebookJsonMappingException;
+import com.restfb.exception.FacebookResponseStatusException;
+
+/**
+ * Generator to convert Facebook error JSON into RestFB Exceptions.
+ *
+ * <p>
+ * Only used to handle errors in case of legacy Facebook access.
+ */
+public interface LegacyFacebookExceptionGenerator {
+
+  /**
+   * If the {@code error_code} JSON field is present, we've got a response status error for this API call. Extracts
+   * relevant information from the JSON and throws an exception which encapsulates it for end-user consumption.
+   *
+   * @param json
+   *          The JSON returned by Facebook in response to an API call.
+   * @param httpStatusCode
+   *          The HTTP status code returned by the server, e.g. 500.
+   * @throws FacebookResponseStatusException
+   *           If the JSON contains an error code.
+   * @throws FacebookJsonMappingException
+   *           If an error occurs while processing the JSON.
+   */
+  void throwLegacyFacebookResponseStatusExceptionIfNecessary(String json, Integer httpStatusCode);
+}
